convert global method to a method in an extension of Date
